Small changes here and there to make it work
Change from int32 to int64 everywhere
Added highestBidder to Outcome

Changes to log
Added longer sleep

+++ server/main.go
@@ -2,28 +2,29 @@ package main
 
 import (
 	"context"
+	"flag"
 	"fmt"
 	"log"
 	"net"
-	"time"
-	"math/rand"
-	"flag"
 	"os"
+	"time"
 
 	gRPC "github.com/Noerklit/DisysAuctionSystem/proto"
 	"google.golang.org/grpc"
 )
 
 type Server struct {
-	gRPC.UnimplementedAuctionSystemServiceServer
-	id               int32
-	port			 string
+	gRPC.UnimplementedAuctionSystemServer
+	id               int64
+	port             string
 	ctx              context.Context
-	highestBid	   	 int32
-	highestBidder 	 string
+	highestBid       int64
+	highestBidder    string
 	AuctionIsOngoing bool
 }
+
 var server *Server
+var counter = 0
 
 var serverName = flag.String("name", "default", "Senders name")
 var serverId = 0
@@ -33,14 +34,19 @@ var _ports [5]string = [5]string{*port, "5401", "5402", "5403", "5404"}
 
 func main() {
 	//Clears the log.txt file when a new server is started
-	if err := os.Truncate("log.txt", 0); err != nil {
-		log.Printf("Failed to truncate: %v", err)
+	if counter == 0 {
+		if err := os.Truncate("log.txt", 0); err != nil {
+			log.Printf("Failed to truncate: %v\n", err)
+			fmt.Printf("Failed to truncate: %v\n", err)
+			counter++
+		}
 	}
 
 	//connect to log file
 	f, err := os.OpenFile("log.txt", os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
 	if err != nil {
-		log.Fatalf("error opening file: %v", err)
+		log.Fatalf("error opening file: %v\n", err)
+		fmt.Printf("error opening file: %v\n", err)
 	}
 	defer f.Close()
 	log.SetOutput(f)
@@ -49,25 +55,26 @@ func main() {
 	go launchServer(_ports[:])
 	go runAuction()
 	for {
-		time.Sleep(time.Second*5)
+		time.Sleep(time.Second * 5)
 	}
 
 }
 
 func runAuction() {
-	durationOfAuction := time.Duration(rand.Intn(45) + 15) * time.Second
+	durationOfAuction := time.Minute * 2
 	time.Sleep(durationOfAuction)
 	server.AuctionIsOngoing = false
-	server.Result()
+	void := &gRPC.Void{}
+	server.Result(server.ctx, void)
 }
 
 func newServer(serverPort string) *Server {
-	server:= &Server{
-		id:               int32(serverId),
-		port:			  serverPort,
+	server := &Server{
+		id:               int64(serverId),
+		port:             serverPort,
 		ctx:              context.Background(),
-		highestBid:		  0,
-		highestBidder:	  "",
+		highestBid:       0,
+		highestBidder:    "",
 		AuctionIsOngoing: true,
 	}
 	serverId++
@@ -77,45 +84,52 @@ func newServer(serverPort string) *Server {
 func launchServer(ports []string) {
 	// Create listener tcp on given port or port 5400
 	log.Printf("Server %s: is trying to create listener on port %s\n", *serverName, ports[0])
+	fmt.Printf("Server %s: is trying to create listener on port %s\n", *serverName, ports[0])
 	list, err := net.Listen("tcp", fmt.Sprintf("localhost:%s", ports[0]))
 	if err != nil {
-		log.Printf("Server %s: Failed to listen on port %s: %v", *serverName, *port, err)
+		log.Printf("Server %s: Failed to listen on port %s: %v\n", *serverName, ports[0], err)
+		fmt.Printf("Server %s: Failed to listen on port %s: %v\n", *serverName, ports[0], err)
 		if len(ports) > 1 {
 			launchServer(ports[1:])
 		} else {
-			log.Fatalf("Server %s: Failed to find open port", *serverName)
+			log.Fatalf("Server %s: Failed to find open port\n", *serverName)
+			fmt.Printf("Server %s: Failed to find open port\n", *serverName)
 		}
+	} else {
+		log.Printf("Server %s: Listening on port %s\n", *serverName, ports[0])
+		fmt.Printf("Server %s: Listening on port %s\n", *serverName, ports[0])
 	}
 
 	var opts []grpc.ServerOption
 	grpcServer := grpc.NewServer(opts...)
 	server = newServer(ports[0])
-	gRPC.RegisterAuctionServiceServer(grpcServer, server)
+	gRPC.RegisterAuctionSystemServer(grpcServer, server)
 
 	if err := grpcServer.Serve(list); err != nil {
-		log.Fatalf("failed to server %v", err)
+		log.Fatalf("failed to server %v\n", err)
+		fmt.Printf("failed to server %v\n", err)
 	}
 }
 
-func(s *Server) Bid(ctx context.Context, amount *gRPC.Amount) (*gRPC.Ack, error){
+func (s *Server) Bid(ctx context.Context, amount *gRPC.Amount) (*gRPC.Ack, error) {
 	if server.AuctionIsOngoing {
 		if amount.Amount > s.highestBid {
 			s.highestBid = amount.Amount
-			s.highestBidder = amount.bidderName
-			fmt.Printf("Server %s: %v has the highest bid of %v\n",*serverName, s.highestBidder, s.highestBid)
-			return &gRPC.Ack{message: "Success"}, nil
+			s.highestBidder = amount.BidderName
+			fmt.Printf("Server %s: %v has the highest bid of %v\n", *serverName, s.highestBidder, s.highestBid)
+			log.Printf("Server %s: %v has the highest bid of %v\n", *serverName, s.highestBidder, s.highestBid)
+			return &gRPC.Ack{Message: "Success"}, nil
 		}
-		return &gRPC.Ack{message: "Fail"}, nil
+		return &gRPC.Ack{Message: "Fail"}, nil
 	} else {
-		return &gRPC.Ack{message: "Exception, the auction is not ongoing, why are you still bidding???"}, nil
+		return &gRPC.Ack{Message: "The auction is not ongoing, why are you still bidding???"}, nil
 	}
 }
 
-
-func(s *Server) Result(ctx context.Context, void *gRPC.Void) (*gRPC.Outcome, error){
+func (s *Server) Result(ctx context.Context, void *gRPC.Void) (*gRPC.Outcome, error) {
 	if s.AuctionIsOngoing {
-		return &gRPC.Result{Outcome: "The auction is still ongoing, the highest bid is currently %d", s.highestBid}, nil
+		return &gRPC.Outcome{Status: false, HighestBid: s.highestBid, HighestBidder: s.highestBidder}, nil
 	} else {
-		return &gRPC.Outcome{Outcome: "The auction is over, the winner is: %v with a bid of: %d!", s.highestBidder, s.highestBid}, nil
+		return &gRPC.Outcome{Status: true, HighestBid: s.highestBid, HighestBidder: s.highestBidder}, nil
 	}
 }
